+++ clients/privacy-center/app/server-environment.ts
@@ -180,6 +180,32 @@ export const validateConfig = (
       };
     }
   }
+
+  const invalidFieldMessages = config.actions.flatMap((action) => {
+    const invalidFields = Object.entries(
+      action.custom_privacy_request_fields || {}
+    )
+      .filter(([, field]) => field.hidden && field.default_value === undefined)
+      .map(([key]) => `'${key}'`);
+
+    return invalidFields.length > 0
+      ? [
+          `${invalidFields.join(", ")} in the action with policy_key '${
+            action.policy_key
+          }'`,
+        ]
+      : [];
+  });
+
+  if (invalidFieldMessages.length > 0) {
+    return {
+      isValid: false,
+      message: `A default_value is required for hidden field(s) ${invalidFieldMessages.join(
+        ", "
+      )}`,
+    };
+  }
+
   return { isValid: true, message: "Config is valid" };
 };